Fix CI by adding a feature spec
The absence of feature specs causes a different, rarely maintained code
branch to be used in lib/tasks/spec.rake. If using the code from the
wagon template for this, any wagon migrations are removed before running
the tests when using `bundle exec rake spec:without_features` (like it
is done on the CI), but it works when migrating the wagon and then just
using plain `rspec` with the same pattern and tag arguments.

In f7ce50a31183c3209464cfaefa5ee06011734f3e, a half-hearted attempt was
made to adjust this so that both code branches would use the same
`RSpec::Core::RakeTask.new`, and this works for running CI, but fails
during assets:precompile when building the Docker images with the
following error:

+ bundle exec rake assets:precompile
rake aborted!
NameError: uninitialized constant RSpec (NameError)

    RSpec::Core::RakeTask.new(:without_features) do |t|
               ^^^^^^^^^^
/app-src/vendor/wagons/hitobito_pfadi_de/lib/tasks/spec.rake:28:in `block in <main>'

So the easiest fix for now, without opening a huge can of worms, is to
just add a feature spec. This should hopefully cause the other, more
frequently used code path to absolve us from these errors.